refract: move ddns module to pkg/
