+++ pkg/ddns/dummy.go
