+++ pkg/ddns/helper.go
